move edge_softmax to functional
