+++ docs/source/api/python/dgl.ops.rst
@@ -236,20 +236,6 @@ The following is an example showing how GSDDMM works:
 
 Like GSpMM, GSDDMM operators support both homogeneous and bipartite graph.
 
-Edge Softmax module
--------------------
-
-DGL also provide framework agnostic edge softmax module which was frequently used in
-GNN-like structures, e.g. 
-`Graph Attention Network <https://arxiv.org/pdf/1710.10903.pdf>`_,
-`Transformer <https://papers.nips.cc/paper/7181-attention-is-all-you-need.pdf>`_,
-`Capsule <https://arxiv.org/pdf/1710.09829.pdf>`_, etc.
-
-.. autosummary::
-    :toctree: ../../generated/
-
-    edge_softmax
-
 Segment Reduce Module
 ---------------------
