+++ python/dgl/nn/pytorch/conv/agnnconv.py
@@ -5,7 +5,7 @@ from torch import nn
 from torch.nn import functional as F
 
 from .... import function as fn
-from ....ops import edge_softmax
+from ...functional import edge_softmax
 from ....base import DGLError
 from ....utils import expand_as_pair
