+++ python/dgl/ops/__init__.py
@@ -1,5 +1,4 @@
 """dgl operator module."""
 from .spmm import *
 from .sddmm import *
-from .edge_softmax import *
 from .segment import *
